Add footer, add eventcard, add FontAwesome icons, removed folders in storybook, modified some existing files, add images
