+++ .storybook/preview-head.html
@@ -1,1 +1,5 @@
 <!-- https://storybook.js.org/docs/configurations/add-custom-head-tags/#docs-content -->
+<script
+  src="https://kit.fontawesome.com/d1f1f68cc2.js"
+  crossorigin="anonymous"
+></script>
