--- src/components/avatar/avatar.css
@@ -1,5 +1,0 @@
-.avatar {
-  border-radius: 50%;
-  height: 100px;
-  width: 100px;
-}
