--- src/components/avatar/avatar.html
@@ -1,1 +1,0 @@
-<img class="avatar" src="../../assets/dog.jpg" alt="Dog" />
