--- src/components/avatar/avatar.stories.js
@@ -1,9 +1,0 @@
-import "./avatar.css";
-import avatar from "./avatar.html";
-
-export default {
-  title: "Components/Avatar",
-  parameters: { layout: "centered" },
-};
-
-export const dog = () => avatar;
