+++ src/components/button/button-home.html
@@ -1,0 +1,4 @@
+<div class="btn-home-grid">
+  <img class="btn-home-img" src="../../assets/home.svg" alt="" />
+  <span class="btn-home-text">Home</span>
+</div>
