--- src/components/button/button-logo.html
@@ -1,1 +1,0 @@
-<button class="btn">🐶</button>
