--- src/components/button/button-text.html
@@ -1,1 +1,0 @@
-<button class="btn">Hello World</button>
