+++ src/components/button/button.stories.js
@@ -1,11 +1,9 @@
 import "./button.css";
-import buttonText from "./button-text.html";
-import buttonLogo from "./button-logo.html";
+import buttonHome from "./button-home.html";
 
 export default {
   title: "Components/Button",
-  parameters: { layout: "centered" },
+  parameters: { layout: "fullscreen" },
 };
 
-export const withText = () => buttonText;
-export const withLogo = () => buttonLogo;
+export const btnhome = () => buttonHome;
