+++ src/components/events/eventcard.html
@@ -1,0 +1,14 @@
+<div class="eventcard-grid">
+  <div class="eventcard-left">
+    <h4>Sports Meet in Galaxy Field</h4>
+    <p><i class="fas fa-calendar-alt"></i> Jan 12, 2019</p>
+    <p>
+      <i class="fas fa-map-marker-alt"></i> Greenfields, Sector 42, Faridabad
+    </p>
+  </div>
+  <img
+    class="eventcard-right-img"
+    src="../../assets/stadium.jpg"
+    alt="Stadium"
+  />
+</div>
