+++ src/components/events/eventcard.stories.js
@@ -1,0 +1,9 @@
+import "./events.css";
+import eventCard from "./eventcard.html";
+
+export default {
+  title: "Components/Events",
+  parameters: { layout: "centered" },
+};
+
+export const eventcard = () => eventCard;
