+++ src/components/footer/footer.html
@@ -1,0 +1,8 @@
+<div class="footer-grid">
+  <div class="btn-home-grid">
+    <img class="btn-home-img" src="../../assets/home.svg" alt="" />
+    <span class="btn-home-text">Home</span>
+  </div>
+  <img class="footer-img" src="../../assets/search.svg" alt="" />
+  <img class="footer-img" src="../../assets/star.svg" alt="" />
+</div>
