+++ src/components/footer/footer.stories.js
@@ -1,0 +1,9 @@
+import "./footer.css";
+import mainFooter from "./footer.html";
+
+export default {
+  title: "Components/Footer",
+  parameters: { layout: "fullscreen" },
+};
+
+export const footer = () => mainFooter;
