--- src/components/header/haeder.stories.js
@@ -1,9 +1,0 @@
-import "./header.css";
-import mainHeader from "./header.html";
-
-export default {
-  title: "Components/Header",
-  parameters: { layout: "fullscreen" },
-};
-
-export const Header = () => mainHeader;
