+++ src/components/header/header.stories.js
@@ -1,0 +1,9 @@
+import "./header.css";
+import mainHeader from "./header.html";
+
+export default {
+  title: "Components/Header",
+  parameters: { layout: "fullscreen" },
+};
+
+export const Header = () => mainHeader;
